+++ src/PhasorFreezer.cpp
@@ -2,13 +2,14 @@
 #include "DSP/Phasors/HCVPhasorEffects.h"
 #include "dsp/digital.hpp"
 
-struct PhasorReset : HCVModule
+struct PhasorFreezer : HCVModule
 {
 	enum ParamIds
 	{
 		PHASE_PARAM,
         PHASECV_PARAM,
         RANGE_PARAM,
+        FREEZE_PARAM,
         RESET_PARAM,
 		NUM_PARAMS
 	};
@@ -16,6 +17,7 @@ struct PhasorReset : HCVModule
 	{
         MAIN_INPUT,
         PHASECV_INPUT,
+        FREEZE_INPUT,
         RESET_INPUT,
 		NUM_INPUTS
 	};
@@ -30,7 +32,7 @@ struct PhasorReset : HCVModule
         NUM_LIGHTS
 	};
 
-	PhasorReset()
+	PhasorFreezer()
 	{
 		config(NUM_PARAMS, NUM_INPUTS, NUM_OUTPUTS, NUM_LIGHTS);
 
@@ -39,27 +41,30 @@ struct PhasorReset : HCVModule
 		configParam(PHASE_PARAM, -5.0f, 5.0f, 0.0f, "Reset Phase");
 		configParam(PHASECV_PARAM, -1.0f, 1.0f, 1.0f, "Reset Phase CV Depth");
         configButton(RESET_PARAM, "Reset");
+        configButton(FREEZE_PARAM, "Freeze");
 
 		configInput(PHASECV_INPUT, "Reset Phase CV");
 
-		configInput(RESET_INPUT, "Reset");
+        configInput(FREEZE_INPUT, "Freeze");
+        configInput(RESET_INPUT, "Reset");
 
 		configInput(MAIN_INPUT, "Phasor");
 		configOutput(MAIN_OUTPUT, "Phasor");
 	}
 
-    HCVPhasorDivMult divmults[16];
+    HCVPhasorFreezer freezers[16];
     dsp::SchmittTrigger inputTriggers[16];
 
 	void process(const ProcessArgs &args) override;
 };
 
 
-void PhasorReset::process(const ProcessArgs &args)
+void PhasorFreezer::process(const ProcessArgs &args)
 {
 	const float shiftKnob = params[PHASE_PARAM].getValue();
 	const float shiftScale = params[PHASECV_INPUT].getValue();
     const float resetButton = params[RESET_PARAM].getValue();
+    const float freezeButton = params[FREEZE_PARAM].getValue();
 
 	int numChannels = setupPolyphonyForAllOutputs();
 
@@ -72,9 +77,11 @@ void PhasorReset::process(const ProcessArgs &args)
         float resetInputs = inputs[RESET_INPUT].getPolyVoltage(i) + resetButton;
         bool resetTrigger = inputTriggers[i].process(resetInputs);
 
-        if(resetTrigger) divmults[i].reset(resetPhase);
+        bool frozen = (inputs[FREEZE_INPUT].getPolyVoltage(i) + freezeButton);
 
-        float output = divmults[i].basicSync(normalizedPhasor);
+        if(resetTrigger) freezers[i].reset(resetPhase);
+
+        float output = freezers[i](normalizedPhasor, frozen);
         outputs[MAIN_OUTPUT].setVoltage(output * HCV_PHZ_UPSCALE, i);
 	}
 
@@ -82,26 +89,30 @@ void PhasorReset::process(const ProcessArgs &args)
 }
 
 
-struct PhasorResetWidget : HCVModuleWidget { PhasorResetWidget(PhasorReset *module); };
+struct PhasorFreezerWidget : HCVModuleWidget { PhasorFreezerWidget(PhasorFreezer *module); };
 
-PhasorResetWidget::PhasorResetWidget(PhasorReset *module)
+PhasorFreezerWidget::PhasorFreezerWidget(PhasorFreezer *module)
 {
-	setSkinPath("res/PhasorReset.svg");
+	setSkinPath("res/PhasorFreezer.svg");
 	initializeWidget(module);
 
 	//////PARAMS//////
-    createParamComboVertical(27, 100, PhasorReset::PHASE_PARAM, PhasorReset::PHASECV_PARAM, PhasorReset::PHASECV_INPUT);
-    createHCVButtonSmall(57, 230, PhasorReset::RESET_PARAM);
+    createParamComboVertical(27, 100, PhasorFreezer::PHASE_PARAM, PhasorFreezer::PHASECV_PARAM, PhasorFreezer::PHASECV_INPUT);
 
 	//////INPUTS//////
     int inputY = 248;
-	createInputPort(13, inputY, PhasorReset::MAIN_INPUT);
-	createInputPort(53, inputY, PhasorReset::RESET_INPUT);
+	createInputPort(13, inputY, PhasorFreezer::FREEZE_INPUT);
+	createInputPort(53, inputY, PhasorFreezer::RESET_INPUT);
+
+    createHCVButtonSmallForJack(13, inputY, PhasorFreezer::FREEZE_PARAM);
+    createHCVButtonSmallForJack(53, inputY, PhasorFreezer::RESET_PARAM);
 
 	//////OUTPUTS//////
-	createOutputPort(33, 300, PhasorReset::MAIN_OUTPUT);
+    int bottomY = 300;
+    createInputPort(13, bottomY, PhasorFreezer::MAIN_INPUT);
+	createOutputPort(53, bottomY, PhasorFreezer::MAIN_OUTPUT);
 
-	createHCVRedLightForJack(33, 300, PhasorReset::PHASOR_LIGHT);
+	createHCVRedLightForJack(53, bottomY, PhasorFreezer::PHASOR_LIGHT);
 }
 
-Model *modelPhasorReset = createModel<PhasorReset, PhasorResetWidget>("PhasorReset");
+Model *modelPhasorFreezer = createModel<PhasorFreezer, PhasorFreezerWidget>("PhasorFreezer");